refactor: share audio waveform view
